+++ testing/dev-tools.mock.ts
@@ -1,0 +1,3 @@
+import { DevToolsExtension } from '@angular-redux/store';
+
+export class MockDevToolsExtension extends DevToolsExtension {}