+++ testing/index.ts
@@ -1,0 +1,3 @@
+export * from './ng-redux-testing.module';
+export * from './dev-tools.mock';
+export * from './ng-redux.mock';